+++ backend/frontend/package.json
@@ -3,10 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@googlemaps/js-api-loader": "^1.15.1",
+    "@material/textfield": "^14.0.0",
+    "@mui/material": "^5.12.2",
+    "@mui/styled-engine": "^5.12.0",
     "@react-google-maps/api": "^2.18.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -16,6 +21,7 @@
     "history": "^5.3.0",
     "jsonwebtoken": "^9.0.0",
     "react": "^18.2.0",
+    "react-animated-heart": "^0.0.8",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-quill": "^2.0.0",
