+++ backend/frontend/src/pages/story/StoryDetails.js
@@ -3,7 +3,11 @@ import axios from 'axios';
 import { useParams,useNavigate  } from 'react-router-dom';
 import styles from './StoryDetails.module.css';
 import { GoogleMap, LoadScriptNext, Marker } from '@react-google-maps/api';
-import withAuth from './authCheck';
+import withAuth from '../../authCheck';
+import { AiOutlineHeart, AiFillHeart } from 'react-icons/ai';
+import Heart from "react-animated-heart";
+
+
 
 
 
@@ -26,6 +30,7 @@ function StoryDetails() {
   const [numLikes, setNumLikes] = useState(0);
   const [liked, setLiked] = useState(false);
   const [userId, setUserId] = useState(null);
+  const [isClick, setClick] = useState(false);
 
 
 
@@ -196,26 +201,13 @@ function StoryDetails() {
               dangerouslySetInnerHTML={{ __html: story.content }}
             />
           <p>{`tags: ${story.story_tags}`}</p>
-          <div>
+          <div className='author'> 
           <span>{numLikes} </span>
           <button
-            className={`heart-button${liked ? ' liked' : ''}`}
             onClick={handleLikeDislike}
             style={{ background: 'none', border: 'none', cursor: 'pointer' }}
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill={liked ? 'red' : 'none'}
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            >
-              <path d="M12 21c-.24 0-.48-.05-.7-.15L2.5 15.5C1.5 14.8 1 13.5 1 12c0-3.3 2.7-6 6-6 2.2 0 4.4 1.7 5 4 .6-2.3 2.8-4 5-4 3.3 0 6 2.7 6 6 0 1.5-.5 2.8-1.5 3.5l-8.8 5.3c-.3.2-.7.2-1 .2z" />
-            </svg>
+             <Heart isClick={liked} onClick={() => setLiked(!liked)}/>
           </button>
         </div>
           {/* <div>
